将picture_dict独立成为一个dict
